Moving header and footer out with the correct name
